First attempt at view recycling
View recycling works, but only for singleton card stacks.  It pushes too
much work and duplication onto the person implementing the card though.
I'll revamp it to be simpler for the API user.  I'm not using multiple
cards per stack though so I probably won't work on that.  This code is
compatible with multiple cards per stack but it just doesn't optimize
for it.